refactor helper methods to separate js file, clear console logs
